requirements.txt
boto3==1.17.85
PyYAML==5.4.1